laplace2d_fpga_experimental app minor changes

+++ apps/c/laplace2d_fpga_experimental/laplace_kernels.h
@@ -1,22 +1,22 @@
 //Kernels for Laplace demo app
 //
-void set_zero(ACC<double> &A) {
+void set_zero(ACC<float> &A) {
   A(0,0) = 0.0;
 }
 
-void copy(ACC<double> &A, const ACC<double> &Anew) {
+void copy(ACC<float> &A, const ACC<float> &Anew) {
   A(0,0) = Anew(0,0);
 }
 
-void left_bndcon(ACC<double> &A, const int *idx) {
+void left_bndcon(ACC<float> &A, const int *idx) {
   A(0,0) = sin(pi * (idx[1]+1) / (jmax+1));
 }
 
-void right_bndcon(ACC<double> &A, const int *idx) {
+void right_bndcon(ACC<float> &A, const int *idx) {
   A(0,0) = sin(pi * (idx[1]+1) / (jmax+1))*exp(-pi);
 }
 
-void apply_stencil(const ACC<double> &A, ACC<double> &Anew) {
+void apply_stencil(const ACC<float> &A, ACC<float> &Anew) {
   Anew(0,0) = 0.25f * ( A(1,0) + A(-1,0)
       + A(0,-1) + A(0,1));
 }
